refactor(users): divide usersController
