--- src/APIs/users/controllers/users-read.controller copy 3.ts
@@ -1,8 +1,0 @@
-import { Controller } from '@nestjs/common';
-import { ApiTags } from '@nestjs/swagger';
-
-@ApiTags('유저 API')
-@Controller('users')
-export class UsersReadController {
-  constructor(private readonly usersService: UsersService) {}
-}
