--- src/APIs/users/controllers/users-update.controller copy.ts
@@ -1,8 +1,0 @@
-import { Controller } from '@nestjs/common';
-import { ApiTags } from '@nestjs/swagger';
-
-@ApiTags('유저 API')
-@Controller('users')
-export class UsersUpdateController {
-  constructor(private readonly usersService: UsersService) {}
-}
